# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Asset Validator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SignupBody {
  fullName: string;
  email: string;
  password: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  user: "user",
} as const;

export interface User {
  id: number;
  fullName: string;
  email: string;
  role: UserRole;
  /** @nullable */
  createdBy?: number | null;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type CreateUserBodyRole =
  (typeof CreateUserBodyRole)[keyof typeof CreateUserBodyRole];

export const CreateUserBodyRole = {
  admin: "admin",
  user: "user",
} as const;

export interface CreateUserBody {
  fullName: string;
  email: string;
  password: string;
  role: CreateUserBodyRole;
}

export type UpdateUserBodyRole =
  (typeof UpdateUserBodyRole)[keyof typeof UpdateUserBodyRole];

export const UpdateUserBodyRole = {
  admin: "admin",
  user: "user",
} as const;

export interface UpdateUserBody {
  fullName?: string;
  email?: string;
  role?: UpdateUserBodyRole;
  /** @nullable */
  password?: string | null;
}

export type ProjectType = (typeof ProjectType)[keyof typeof ProjectType];

export const ProjectType = {
  image: "image",
  text: "text",
  audio: "audio",
  video: "video",
} as const;

export interface Project {
  id: number;
  name: string;
  type: ProjectType;
  userId: number;
  createdAt: string;
}

export type ProjectWithStatsType =
  (typeof ProjectWithStatsType)[keyof typeof ProjectWithStatsType];

export const ProjectWithStatsType = {
  image: "image",
  text: "text",
  audio: "audio",
  video: "video",
} as const;

export interface ProjectWithStats {
  id: number;
  name: string;
  type: ProjectWithStatsType;
  userId: number;
  createdAt: string;
  totalAssets: number;
  validatedAssets: number;
  passCount: number;
  passPercent: number;
  totalTokens: number;
  totalCost: number;
  avgLatency: number;
}

export type CreateProjectBodyType =
  (typeof CreateProjectBodyType)[keyof typeof CreateProjectBodyType];

export const CreateProjectBodyType = {
  image: "image",
  text: "text",
  audio: "audio",
  video: "video",
} as const;

export interface CreateProjectBody {
  name: string;
  type: CreateProjectBodyType;
}

export interface ProjectConfig {
  id: number;
  projectId: number;
  validationRules: string;
  enableDuplicationCheck: boolean;
  enablePIIValidation: boolean;
  enableBlurCheck: boolean;
  updatedAt: string;
}

export interface UpsertProjectConfigBody {
  validationRules?: string;
  enableDuplicationCheck?: boolean;
  enablePIIValidation?: boolean;
  enableBlurCheck?: boolean;
}

export type ValidateAssetBodyAssetType =
  (typeof ValidateAssetBodyAssetType)[keyof typeof ValidateAssetBodyAssetType];

export const ValidateAssetBodyAssetType = {
  image: "image",
  text: "text",
  audio: "audio",
  video: "video",
} as const;

export interface ValidateAssetBody {
  projectId: number;
  assetName: string;
  assetContent: string;
  assetType: ValidateAssetBodyAssetType;
  /** @nullable */
  validationRules?: string | null;
  /** @nullable */
  promptOverride?: string | null;
  /** @nullable */
  model?: string | null;
}

export type ValidationResultStatus =
  (typeof ValidationResultStatus)[keyof typeof ValidationResultStatus];

export const ValidationResultStatus = {
  PASS: "PASS",
  FAIL: "FAIL",
} as const;

export type ValidationResultPreCheckResults = {
  /** @nullable */
  blurScore?: number | null;
  isDuplicate: boolean;
  piiDetected: boolean;
  piiItems: string[];
};

export interface ValidationResult {
  id: number;
  status: ValidationResultStatus;
  reasons: string[];
  confidence: number;
  tokensUsed: number;
  cost: number;
  latency: number;
  rawResponse: string;
  preCheckResults: ValidationResultPreCheckResults;
}

export type AssetValidationValidationResult =
  (typeof AssetValidationValidationResult)[keyof typeof AssetValidationValidationResult];

export const AssetValidationValidationResult = {
  PASS: "PASS",
  FAIL: "FAIL",
} as const;

export interface AssetValidation {
  id: number;
  projectId: number;
  projectName: string;
  assetName: string;
  validationResult: AssetValidationValidationResult;
  reasons: string[];
  tokensUsed: number;
  cost: number;
  latency: number;
  confidence: number;
  createdAt: string;
}

export type ObservabilityResponseMetrics = {
  totalCost: number;
  totalTokens: number;
  avgLatency: number;
  totalRecords: number;
  passCount: number;
  failCount: number;
};

export interface ObservabilityResponse {
  records: AssetValidation[];
  metrics: ObservabilityResponseMetrics;
}

export type SystemPromptModality =
  (typeof SystemPromptModality)[keyof typeof SystemPromptModality];

export const SystemPromptModality = {
  image: "image",
  text: "text",
  audio: "audio",
  video: "video",
} as const;

export interface SystemPrompt {
  id: number;
  modality: SystemPromptModality;
  version: string;
  prompt: string;
  isActive: boolean;
  createdAt: string;
}

export type CreatePromptBodyModality =
  (typeof CreatePromptBodyModality)[keyof typeof CreatePromptBodyModality];

export const CreatePromptBodyModality = {
  image: "image",
  text: "text",
  audio: "audio",
  video: "video",
} as const;

export interface CreatePromptBody {
  modality: CreatePromptBodyModality;
  version: string;
  prompt: string;
  isActive?: boolean;
}

export type UpdatePromptBodyModality =
  (typeof UpdatePromptBodyModality)[keyof typeof UpdatePromptBodyModality];

export const UpdatePromptBodyModality = {
  image: "image",
  text: "text",
  audio: "audio",
  video: "video",
} as const;

export interface UpdatePromptBody {
  modality?: UpdatePromptBodyModality;
  version?: string;
  prompt?: string;
  isActive?: boolean;
}

export type ListValidationsParams = {
  /**
   * @nullable
   */
  projectId?: number | null;
  /**
   * @nullable
   */
  result?: ListValidationsResult;
};

export type ListValidationsResult =
  | (typeof ListValidationsResult)[keyof typeof ListValidationsResult]
  | null;

export const ListValidationsResult = {
  PASS: "PASS",
  FAIL: "FAIL",
} as const;

export type GetObservabilityParams = {
  /**
   * @nullable
   */
  projectId?: number | null;
  /**
   * @nullable
   */
  result?: GetObservabilityResult;
};

export type GetObservabilityResult =
  | (typeof GetObservabilityResult)[keyof typeof GetObservabilityResult]
  | null;

export const GetObservabilityResult = {
  PASS: "PASS",
  FAIL: "FAIL",
} as const;