# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Asset Validator API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Sign up
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project with stats
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project with stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectWithStats"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{id}/config:
    get:
      operationId: getProjectConfig
      tags: [config]
      summary: Get project configuration
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectConfig"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: upsertProjectConfig
      tags: [config]
      summary: Create or update project configuration
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertProjectConfigBody"
      responses:
        "200":
          description: Config updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectConfig"
...
  /prompts:
    get:
      operationId: listPrompts
      tags: [prompts]
      summary: List system prompts
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of prompts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SystemPrompt"
    post:
      operationId: createPrompt
      tags: [prompts]
      summary: Create a system prompt (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePromptBody"
      responses:
        "201":
          description: Prompt created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemPrompt"
  /prompts/{id}:
    put:
      operationId: updatePrompt
      tags: [prompts]
      summary: Update a system prompt (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePromptBody"
      responses:
        "200":
          description: Prompt updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemPrompt"
    delete:
      operationId: deletePrompt
      tags: [prompts]
      summary: Delete a system prompt (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Prompt deleted
  /prompts/{id}/activate:
    post:
      operationId: activatePrompt
      tags: [prompts]
      summary: Set a prompt as active for its modality (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Prompt activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemPrompt"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SignupBody:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
        password:
          type: string
      required:
        - fullName
        - email
        - password
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, user]
        createdBy:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - fullName
        - email
        - role
        - createdAt
...
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [image, text, audio, video]
        userId:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - type
        - userId
        - createdAt
    ProjectWithStats:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [image, text, audio, video]
        userId:
          type: integer
        createdAt:
          type: string
          format: date-time
        totalAssets:
          type: integer
        validatedAssets:
          type: integer
        passCount:
          type: integer
        passPercent:
          type: number
        totalTokens:
          type: integer
        totalCost:
          type: number
        avgLatency:
          type: number
      required:
        - id
        - name
        - type
        - userId
        - createdAt
        - totalAssets
        - validatedAssets
        - passCount
        - passPercent
        - totalTokens
        - totalCost
        - avgLatency
    CreateProjectBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [image, text, audio, video]
      required:
        - name
        - type
    ProjectConfig:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        validationRules:
          type: string
        enableDuplicationCheck:
          type: boolean
        enablePIIValidation:
          type: boolean
        enableBlurCheck:
          type: boolean
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - validationRules
        - enableDuplicationCheck
        - enablePIIValidation
        - enableBlurCheck
        - updatedAt
    UpsertProjectConfigBody:
      type: object
      properties:
        validationRules:
          type: string
        enableDuplicationCheck:
          type: boolean
        enablePIIValidation:
          type: boolean
        enableBlurCheck:
          type: boolean
...
    SystemPrompt:
      type: object
      properties:
        id:
          type: integer
        modality:
          type: string
          enum: [image, text, audio, video]
        version:
          type: string
        prompt:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - modality
        - version
        - prompt
        - isActive
        - createdAt
    CreatePromptBody:
      type: object
      properties:
        modality:
          type: string
          enum: [image, text, audio, video]
        version:
          type: string
        prompt:
          type: string
        isActive:
          type: boolean
      required:
        - modality
        - version
        - prompt
    UpdatePromptBody:
      type: object
      properties:
        modality:
          type: string
          enum: [image, text, audio, video]
        version:
          type: string
        prompt:
          type: string
        isActive:
          type: boolean