# lib/api-zod/src/generated/types/projectWithStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Asset Validator API
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectWithStatsType } from "./projectWithStatsType";

export interface ProjectWithStats {
  id: number;
  name: string;
  type: ProjectWithStatsType;
  userId: number;
  createdAt: Date;
  totalAssets: number;
  validatedAssets: number;
  passCount: number;
  passPercent: number;
  totalTokens: number;
  totalCost: number;
  avgLatency: number;
}