# lib/api-zod/src/generated/types/systemPrompt.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Asset Validator API
 * OpenAPI spec version: 0.1.0
 */
import type { SystemPromptModality } from "./systemPromptModality";

export interface SystemPrompt {
  id: number;
  modality: SystemPromptModality;
  version: string;
  prompt: string;
  isActive: boolean;
  createdAt: Date;
}