# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Asset Validator API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: number;
  fullName: string;
  email: string;
  role: UserRole;
  /** @nullable */
  createdBy?: number | null;
  createdAt: Date;
}